pyproject.toml
[project]
name = "sony-restock-checker"
version = "0.1.0"
description = "Sony restock checker"
requires-python = ">=3.12"
dependencies = [
    "playwright==1.48.0",
    "aiohttp==3.11.11",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["main"]

[tool.uv]
dev-dependencies = []

